+++ OgreMain/src/OgreCompositorManager.cpp
@@ -222,7 +222,7 @@ void CompositorManager::_reconstructAllCompositorResources()
 //---------------------------------------------------------------------
 TexturePtr CompositorManager::getPooledTexture(const String& name, 
     const String& localName,
-    size_t w, size_t h, PixelFormat f, uint aa, const String& aaHint, bool srgb, 
+    uint32 w, uint32 h, PixelFormat f, uint aa, const String& aaHint, bool srgb,
     CompositorManager::UniqueTextureSet& texturesAssigned, 
     CompositorInstance* inst, CompositionTechnique::TextureScope scope)
 {
@@ -310,7 +310,7 @@ TexturePtr CompositorManager::getPooledTexture(const String& name,
         ret = TextureManager::getSingleton().createManual(
             name, 
             ResourceGroupManager::INTERNAL_RESOURCE_GROUP_NAME, TEX_TYPE_2D, 
-            (uint)w, (uint)h, 0, f, TU_RENDERTARGET, 0,
+            w, h, 0, f, TU_RENDERTARGET, 0,
             srgb, aa, aaHint); 
 
         texList->push_back(ret);
